gemini/requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain-openai
langchain-community
